feat: generic ShadowScrollArea component
This creates a generic ShadowScrollArea component we can reuse eleswhere
